- Implementado mecanismo de seguridad en varias apis - Implementado servicio de logout - Implementada la seguridad basica del PDO - Implementados varios fixes de formato

+++ server/test/test-pdo_group.js
@@ -9,9 +9,22 @@ var PdoGroup = mongoose.model('PdoGroup');
 var sprintf = require("sprintf-js").sprintf;
 var q = require('q');
 describe('Test de grupos de PDOs', function() {
-    var school, program, course, created_pdo = [],
-        pdo_to_add_later,
-        pdo_group_created, created_comment_id;
+    var school;
+    var program; 
+    var course;
+    var created_pdo = [];
+    var pdo_to_add_later;
+    var pdo_group_created;
+    var created_comment_id;
+    var other_school;
+    var pdo_other_school;
+    var token = {
+        'Authorization': ''
+    };
+    var test_user = {
+        username: 'joaquin.fernandez+schoolAdmin@edu.uah.es',
+        password: '1234'
+    };
     before(function(done) {
         var prom_array = [];
         mongoose.connect(config.db.mongodb, function(err) {
@@ -34,11 +47,19 @@ describe('Test de grupos de PDOs', function() {
                 }).exec(),
                 Course.findOne({
                     code: '570000'
-                }).exec()
+                }).exec(),
+                School.findOne({
+                    schoolname: 'CCSS'
+                }).exec(),
+                request(url).
+                post('/api/v1/login').
+                send(test_user)
             ]).then(function(results) {
                 program = results[0];
                 school = results[1];
                 course = results[2];
+                other_school = results[3];
+                token.Authorization = results[4].body.token;
                 var pdo;
                 for (var i = 0; i < 3; i++) {
                     pdo = {};
@@ -52,19 +73,41 @@ describe('Test de grupos de PDOs', function() {
                     pdo.course = course;
                     pdo.title = sprintf("TITULO DEL PDO %d", i);
                     pdo.text = sprintf("TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO %d", i);
-                    prom_array.push(request(url).post('/api/v1/pdo').send(pdo).expect('Content-Type', /json/).expect(200));
+                    prom_array.push(
+                        request(url).
+                        post('/api/v1/pdo').
+                        send(pdo).
+                        expect('Content-Type', /json/).
+                        expect(200));
                 }
+                for (var i = 3; i < 5; i++) {
+                    pdo = {};
+                    pdo.name = sprintf('PDOCCSS Nombre%d', i);
+                    pdo.surname = sprintf('PDOCCSS Apellido%d', i);
+                    pdo.num_id = '123456789A';
+                    pdo.email = sprintf('jfcampo+%d@gmail.com', i);
+                    pdo.phone = '654654654';
+                    pdo.school = other_school;
+                    pdo.program = program;
+                    pdo.course = course;
+                    pdo.title = sprintf("PDOCCSS TITULO DEL PDO %d", i);
+                    pdo.text = sprintf("PDOCCSS TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO TEXTO %d", i);
+                    prom_array.push(
+                        request(url).
+                        post('/api/v1/pdo').
+                        send(pdo).
+                        expect('Content-Type', /json/).
+                        expect(200));
+                };
+                
                 q.all(prom_array).then(function(results) {
-                    for (var i = results.length - 1; i >= 0; i--) {
-                        if (i !== 1) {
-                            created_pdo.push(results[i].body);
-                        } else {
-                            pdo_to_add_later = results[i].body;
-                        }
-                    }
+                    created_pdo = results.slice(0,2).map(function(e){return e.body;});
+                    pdo_to_add_later = results[2].body;
+                    pdo_other_school = results.slice(3).map(function(e){return e.body;});
                     done();
                 }).
                 catch (function(reason) {
+                    console.log('Reason:',reason);
                     done(reason);
                 });
             }).
@@ -80,11 +123,18 @@ describe('Test de grupos de PDOs', function() {
         pdo_group.pdos = JSON.stringify(created_pdo.map(function(pdo) {
             return pdo._id;
         }));
-        prom = request(url).post('/api/v1/pdo_group').send(pdo_group).expect('Content-Type', /json/).expect(200).then(function(result) {
+        prom = request(url).
+        post('/api/v1/pdo_group').
+        set(token).
+        send(pdo_group).
+        expect('Content-Type', /json/).
+        expect(200).
+        then(function(result) {
             pdo_group_created = result.body;
             pdo_group_created.status.should.equal('STATUS_PENDING');
             pdo_group_created.title.should.equal(pdo_group.title);
-            pdo_group_created.should.have.property('pdos').with.length(2);
+            pdo_group_created.should.have.property('pdos').with.length(created_pdo.length);
+            console.log('CreatedPdo', created_pdo);
             pdo_group_created.pdos[0].should.equal(created_pdo[0]._id);
             done();
         }, function(reason) {
@@ -95,7 +145,11 @@ describe('Test de grupos de PDOs', function() {
     it('should add a pdo to a group', function(done) {
         var prom, payload = {};
         payload.pdos = pdo_to_add_later._id;
-        prom = request(url).put('/api/v1/pdo_group/' + pdo_group_created._id + '/pdo').send(payload).expect('Content-Type', /json/).expect(200).then(function(result) {
+        prom = request(url).
+        put('/api/v1/pdo_group/' + pdo_group_created._id + '/pdo').
+        set(token).
+        send(payload).
+        expect('Content-Type', /json/).expect(200).then(function(result) {
             pdo_group_created = result.body;
             pdo_group_created.status.should.equal('STATUS_PENDING');
             pdo_group_created.should.have.property('pdos').with.length(3);
@@ -144,7 +198,7 @@ describe('Test de grupos de PDOs', function() {
             done(reason);
         });
     });*/
-    it('an added pdo should have a pdo_group', function(done) {
+    it('a grouped pdo should have a pdo_group', function(done) {
         Pdo.findById(pdo_to_add_later._id, function(err, pdo) {
             if (err) {
                 done(err);
@@ -160,7 +214,10 @@ describe('Test de grupos de PDOs', function() {
     it('should remove a pdo from a group', function(done) {
         var prom, payload = {};
         payload.pdo_id = pdo_to_add_later._id;
-        prom = request(url).delete('/api/v1/pdo_group/' + pdo_group_created._id + '/pdo').send(payload).expect('Content-Type', /json/).expect(200);
+        prom = request(url).
+        delete('/api/v1/pdo_group/' + pdo_group_created._id + '/pdo').
+        set(token).
+        send(payload).expect('Content-Type', /json/).expect(200);
         prom.then(function(result) {
             var pdo_group_edited = result.body;
             pdo_group_edited.should.have.property('pdos');
@@ -174,7 +231,7 @@ describe('Test de grupos de PDOs', function() {
         });
         return prom;
     });
-    it('a pdo should not have a pdo_group', function(done) {
+    it('a degrouped pdo should not have a pdo_group', function(done) {
         var pdo_deleted_from_group = pdo_to_add_later;
         Pdo.findById(pdo_deleted_from_group._id, function(err, found_pdo) {
             if (err) {
@@ -190,7 +247,10 @@ describe('Test de grupos de PDOs', function() {
     it('should not add a comment if it doesn\'t have text', function(done) {
         var bad_comment = {};
         bad_comment.title = 'Titulo del comentario ' + faker.random.number();
-        request(url).put('/api/v1/pdo_group/' + pdo_group_created._id + '/comment').send(bad_comment).expect('Content-Type', /json/).expect(400).then(function(res) {
+        request(url).
+        put('/api/v1/pdo_group/' + pdo_group_created._id + '/comment').
+        set(token).
+        send(bad_comment).expect('Content-Type', /json/).expect(400).then(function(res) {
             //console.log(res.body);
             res.body.errors.text.type.should.equal('required');
             done();
@@ -204,7 +264,10 @@ describe('Test de grupos de PDOs', function() {
         var bad_comment = {};
         bad_comment.title = 'Titulo del comentario ' + faker.random.number();
         bad_comment.text = ' ';
-        request(url).put('/api/v1/pdo_group/' + pdo_group_created._id + '/comment').send(bad_comment).expect('Content-Type', /json/).expect(400).then(function(res) {
+        request(url).
+        put('/api/v1/pdo_group/' + pdo_group_created._id + '/comment').
+        set(token).
+        send(bad_comment).expect('Content-Type', /json/).expect(400).then(function(res) {
             res.body.errors.text.message.should.equal('Text length should be between 50 and 3500 chars');
             done();
         }).
@@ -219,9 +282,8 @@ describe('Test de grupos de PDOs', function() {
         bad_comment.text = faker.lorem.paragraphs();
         request(url).
         put('/api/v1/pdo_group/' + pdo_group_created._id + '/comment').
-        send(bad_comment)
-        .expect('Content-Type', /json/)
-        .expect(400).then(function(res) {
+        set(token).
+        send(bad_comment).expect('Content-Type', /json/).expect(400).then(function(res) {
             res.body.errors.title.type.should.equal('required');
             done();
         }).
@@ -236,6 +298,7 @@ describe('Test de grupos de PDOs', function() {
         bad_comment.text = faker.lorem.paragraphs();
         request(url).
         put('/api/v1/pdo_group/' + pdo_group_created._id + '/comment').
+        set(token).
         send(bad_comment).
         expect('Content-Type', /json/).
         expect(400).
@@ -248,12 +311,37 @@ describe('Test de grupos de PDOs', function() {
             done(reason);
         });
     });
+    it('should not group pdos from another school', function (done) {
+        var pdo_group = {}, pdo_group_creado, prom;
+        pdo_group.title = 'Titulo del grupo de pdo - fail';
+        pdo_group.summary = 'Cosas muy malas han pasado -fail';
+        pdo_group.pdos = JSON.stringify(pdo_other_school.map(function(pdo) {
+            return pdo._id;
+        }));
+        prom = request(url).
+        post('/api/v1/pdo_group').
+        set(token).
+        send(pdo_group).
+        expect('Content-Type', /json/).
+        expect(401).
+        then(function(result) {
+            done();
+        }, function(reason) {
+            done(reason);
+        });
+        return prom.promise;
+    });
     after(function(done) {
         //Limpieza
+        if (typeof pdo_group_created === 'undefined') {
+            done();
+            return;
+        }
         created_pdo.push(pdo_to_add_later);
         q.all(created_pdo.map(function(pdo) {
             return Pdo.findByIdAndRemove(pdo._id).exec();
-        }).push(PdoGroup.findByIdAndRemove(pdo_group_created._id).exec())).then(function(result) {
+        }).push(PdoGroup.findByIdAndRemove(pdo_group_created._id).exec())).
+        then(function(result) {
             mongoose.connection.close();
             done();
         }).
